requirements.txt
altgraph==0.17.2
asn1crypto==1.5.1        
attrs==21.4.0
auto-py-to-exe==2.20.1   
bottle==0.12.21
bottle-websocket==0.2.9  
certifi==2022.12.7       
cffi==1.15.0
charset-normalizer==2.1.1
cryptography==38.0.4     
Eel==0.14.0
filelock==3.9.0
flatten-json==0.1.13     
format-sql==0.12.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.4
numpy==1.22.4
oscrypto==1.3.0
pandas==1.4.2
pefile==2022.5.30
pycparser==2.21
pycryptodomex==3.16.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.6
PyJWT==2.6.0
pyodbc==4.0.35
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.0
requests==2.28.1
six==1.16.0
snowflake-connector-python==2.8.3
sql-formatter==0.6.2
sqlparse==0.4.2
typing_extensions==4.4.0
urllib3==1.26.13
whichcraft==0.6.1
XlsxWriter==3.0.3
zope.event==4.5.0
zope.interface==5.4.0